* Implemented both versions of setMap * Bugfix, tile map height was not calculated correctly at initialization * Use both versions of setMap in demo02

+++ source/rage.h
@@ -89,15 +89,13 @@ class Rage
   const char *getErrorString();
   int redraw();
   int selectOnTop(Engine e);
-  int setupBackground(Engine e, int layer, int tileWidth, int tileHeight);
+  int setupBackground(Engine e, u16 layer, u16 tileWidth, u16 tileHeight);
   int loadTileSet(Engine e, TileSetDefinition *def);
   int unloadTileSet(Engine e, u16 tileSet);
   int unloadAllTileSets(Engine e);
-  int setTile(Engine e, int layer, u16 x, u16 y, u16 tileSet, u16 tile);
-  int setMap(Engine e, int layer, Tile *map);
-  int setMap(Engine e, int layer, int tileSet, u16 *map);
-
-  // DEBUG
+  int setTile(Engine e, u16 layer, u16 x, u16 y, u16 tileSet, u16 tile);
+  int setMap(Engine e, u16 layer, Tile *map);
+  int setMap(Engine e, u16 layer, u16 tileSet, u16 *map);
   void listFreeBlocks();
 };
